requirements.txt
# Animus CLI Dependencies

# Core packages
argparse>=1.4.0
colorama>=0.4.6   # For colored terminal output
pyreadline3>=3.4.0;platform_system=="Windows"  # For command history support on Windows systems
tabulate>=0.9.0   # For table formatting

# Google Gemini integration
google-generativeai
python-dotenv
